DB-11919 cleanup.
