+++ db-engine/src/main/java/com/splicemachine/db/iapi/sql/compile/Optimizer.java
@@ -425,5 +425,5 @@ public interface Optimizer{
 
     default boolean isMemPlatform() { return false; };
 
-    JoinCostEstimationModel getCurrentJoinCostEstimationModel();
+    JoinCostEstimationModel getJoinCostEstimationModel();
 }
