+++ db-engine/src/main/java/com/splicemachine/db/impl/ast/RSUtils.java
@@ -285,8 +285,7 @@ public class RSUtils {
     }
 
     public static boolean isMSJ(AccessPath ap) {
-        return (ap != null && (ap.getJoinStrategy().getJoinStrategyType().equals(JoinStrategy.JoinStrategyType.MERGE_SORT) ||
-                (ap.getJoinStrategy().getJoinStrategyType().equals(JoinStrategy.JoinStrategyType.HALF_MERGE_SORT))));
+        return (ap != null && ap.getJoinStrategy().getJoinStrategyType().equals(JoinStrategy.JoinStrategyType.MERGE_SORT));
     }
 
     public static boolean isMJ(AccessPath ap) {
